refactor: create a super class for base page
