+++ packages/types/src/page/extended.d.ts
@@ -1,0 +1,1 @@
+export type TPageExtended = {};
