+++ packages/types/src/page/index.d.ts
@@ -1,0 +1,2 @@
+export * from "./core";
+export * from "./extended";
