+++ packages/types/src/search.d.ts
@@ -1,7 +1,7 @@
 import { ICycle } from "./cycle";
 import { TIssue } from "./issues/issue";
 import { IModule } from "./module";
-import { TPage } from "./pages";
+import { TPage } from "./page/core";
 import { IProject } from "./project";
 import { IUser } from "./users";
 import { IWorkspace } from "./workspace";
